refactor(Ticket): Serializable 제거, Position 테스트 코드 작성, 테스트코드 패키지 구조 정리
