+++ domain/src/main/java/woowacourse/movie/ticket/Position.kt
@@ -5,7 +5,7 @@ data class Position(
     val column: Int,
 ) {
     init {
-        require(column >= 0) { "좌석의 행은 음수일 수 없습니다. 현재 열: $column" }
-        require(row >= 0) { "좌석의 열은 음수일 수 없습니다. 현재 열: $column" }
+        require(row >= 0) { "좌석의 행은 음수일 수 없습니다. 현재 행: $row" }
+        require(column >= 0) { "좌석의 열은 음수일 수 없습니다. 현재 열: $column" }
     }
 }
